backend/requirements.txt
firebase-functions>=0.1.0
firebase-admin>=6.5.0
google-genai>=1.10.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0
